[WIP][Improve] reorganize the menus
