+++ streampark-console/streampark-console-webapp/src/hooks/web/useLog.ts
