+++ streampark-console/streampark-console-webapp/src/locales/lang/en/menu.ts
@@ -23,17 +23,19 @@ export default {
     tokenManagement: 'Token Management',
     teamManagement: 'Team Management',
     memberManagement: 'Member Management',
-    project: 'Project',
-    application: 'Application',
-    variable: 'Variable',
-    resource: 'Resource',
+    project: 'Projects',
+    application: 'Applications',
+    variable: 'Variables',
+    resource: 'Resources',
     setting: 'Settings',
   },
-  setting: {
-    system: 'System Setting',
-    alarm: 'Alarm Setting',
+  flink: {
     flinkHome: 'Flink Home',
-    flinkCluster: 'Flink Cluster',
+    flinkCluster: 'Clusters',
+  },
+  setting: {
+    system: 'Environment Settings',
+    alarm: 'Alarm Settings',
     externalLink: 'External Link',
     yarnQueue: 'Yarn Queue',
     flinkGateway: 'FLink Gateway',
