+++ streampark-console/streampark-console-webapp/src/locales/lang/zh-CN/menu.ts
@@ -23,17 +23,19 @@ export default {
     tokenManagement: 'Token 管理',
     teamManagement: '团队管理',
     memberManagement: '成员管理',
-    project: '项目管理',
+    project: '项目',
     application: '作业管理',
     variable: '变量管理',
     resource: '资源管理',
     setting: '设置中心',
   },
+  flink: {
+    flinkHome: 'Flink版本',
+    flinkCluster: '集群',
+  },
   setting: {
     system: '系统设置',
     alarm: '告警设置',
-    flinkHome: 'Flink 版本',
-    flinkCluster: 'Flink 集群',
     externalLink: '扩展链接',
     yarnQueue: 'Yarn 队列',
     flinkGateway: 'FLink 网关',
