+++ streampark-console/streampark-console-webapp/src/views/project/View.vue
@@ -121,7 +121,7 @@
       let projectDataSource = ref<Array<ProjectRecord>>([]);
 
       function onAdd() {
-        go(`/flink/project/add`);
+        go(`/project/add`);
       }
 
       function handleSearch(value: string) {
